feat(runtime): add setup option for agnostic runtime support
By moving the runtime to a config option which is set during setup the
runtime parameter is not necessary anymore for each LSP `cmd`.

Also rename the global configuration variable to something plugin
related to make it obvious which plugin is responsible for that variable
and to avoid possible name collisions.