Feature: Add new backend: nix
This PR introduces a new optional backend to pull the language servers
from: nix.
Nix does not necessarily fit the containerization theme of this plugin,
but provides a very similar user experience to the docker backend.
It also cuts down on the (albeit very slim) containerization overhead.

The nix backend provides the same interface as the docker/podman
approach in that it can pull/build and delete language servers dynamically
to save on disk space.