feat: add `ts_ls` language server support
nvim-lsp has updated the name of `tsserver` to `ts_ls` for
disambiguation purposes. The underlying language server is still
the same, so support should be as simple as adding it to the list
of supported servers with the same image as `tsserver`. Lmk if I
need to do anything else to get this across